Add indexing endpoint
When I tried to use RPC call to index my text, I didn't know what is the indexing endpoint. I searched it on the portal and found nothing and then I searched your github source code to get the answer. I think it will be more convenient and user-friendly if you can put the endpoint on the document page.

Signed-off-by: Jean.T <yt2101@nyu.edu>