Refocus overview.md with active voice
Most documentation focuses on what the user should do to accomplish something.

This page is different. It establishes some commitments. You may want to review it with legal experts. But if I understand correctly, it a list of standard Vectara design practices.

With active voice, this page would establish a relationship, with a commitment from the people of Vectara to their customers. 

I have some open questions, such as whether you want to specify a TLS version, since 1.0 and 1.1 are now obsolete. Would the typical reader assume that you're using the latest version of TLS (1.3) or what might be the "mainstream" version (1.2)? Or does Vectara accommodate both?

Signed-off-by: Mike Jang <michael@linuxexam.com>