Update Quick Start (#82)
* Add Quick Start Tutorial

Created a Quick Start for the Getting Started section. This first quick start is for the console UI.

* Update quickstart.md

Made some updates based on feedback

* Update quickstart.md

Minor updates to the new quick start

* Update quickstart and graphics

I updated the quick start tutorial to be more focused and fun on a sample document. I also updated the home page to link to the tutorial.

* Update quickstart

Made some updates based on suggestions

---------

Signed-off-by: Paul Wozniczka <25128922+pwoznic@users.noreply.github.com>