Improve indexing topic - WIP
Improving this topic with the goal to make all API Definition topics more standardized. I plan to move the conceptual information higher in the topic (e.g. describing objects and sections) and minimize the amount of code samples in these API topics and snippets because having so many is hard to maintain. Instead, we'll link to the API playground and proto files.