Improve the Index API Definition topic (#208)
* Add topic to help guide users on which indexing API to choose

We have been getting feedback that we need a topic or landing page that helps users understand our different indexing APIs at a higher level. - WIP

* Update indexing section

WIP

* WIP

* Improve structure of the ADI Definition topics

Updating this topic and working on a better format for the long-form topics overall - WIP

* Improve indexing topic - WIP

Improving this topic with the goal to make all API Definition topics more standardized. I plan to move the conceptual information higher in the topic (e.g. describing objects and sections) and minimize the amount of code samples in these API topics and snippets because having so many is hard to maintain. Instead, we'll link to the API playground and proto files.

* Update long-form indexing topic

Moving things around to improve the flow of this topic. For example, start by describing the endpoint without being too REST or gRPC specific. I also changed the API Playground and examples links to be a tip (stands out better). After going through the conceptual information (e.g. defining document and section), I provide some REST and gRPC examples. Still WIP as I might remove the examples and point to the playground and proto definitions.

* Update Low-level Indexing topic

WIP

* Continue updating indexing long-form topics

WIP

* Update high-lvl topic

* WIP

* Updates based on feedback

* Update indexing topics and sidebar

* Update Delete Doc with new topic template

* Fix URL

* Update indexing.md