Update part.is_title field description (#216)
* Update part.is_title field description

Updated part.is_title field to include when it's not equal to true (<>). This is useful when documents do not have titles and you want to search for sections without titles. Setting to false will not work because there are cases where this field is unset.

* Update ootb-metadata-filters.md

* Update ootb-metadata-filters.md

* Update ootb-metadata-filters.md

Going through issues and updated part.is_title to be more clear